Failover Clustering/Deploy: Set-ClusterQuorum URL
Description:

As reported in issue ticket #4454 (Set-ClusterQuorum link is outdated),
the link URL for Set-ClusterQuorum is outdated by currently using an
old TechNet link permanently redirected to docs.microsoft.com and that
the page belongs to the "Windows Server 2008 R2 and 2008" subdirectory
of the Previous Versions hierarchy. Better to use a more updated page.

Thanks to grayjeremy (Jeremy Gray) for pointing out this issue.

Changes proposed:
- Replace the old TechNet URL with the new PowerShell page URL
- Typo correction: "Configue" -> Configure
- Whitespace and codestyle adjustments:
    - Remove redundant end-of-line blanks (53 lines)
    - Add a blank line between metadata and the page title
    - Normalize H2/H3/H4 headings by adding blank line after (10 lines)
    - Standardize on hyphens as bullet point list indicators (3 lines)
    - Standardize on 1 blank line above image link (1 line)

Additional note: When reviewing this PR source, it is recommended to
use the "Hide whitespace changes" feature for readability, because git
reacts to the changes in number of lines and messes up the line order
between old and new lines, mostly because of the added blank lines.

Ticket closure or reference:

Closes #4454