ADDS/Capacity Planning: remove forgotten comment
Description:

As reported in the issue tickets #4339 (Comment in Table) and #4470
("are you sure ???"), there is a curious comment in the RAM Calculation
summary example table after the Database (Global Catalog) | 8.5 GB
recommendation saying exactly that: `are you sure ???`

Thanks to Shawn Barton (shawnbarton) for originally reporting this in
the German localization part of this repo, and to Merola (Merola132)
for reminding us about it again today in the default repository.

Thanks also to srvbpigh, Cecilia Acevedo (CeciAc) and Ramiro Del Greco
(v-radelg) for moving the localization ticket to the source repository.

Changes proposed:
- Remove the comment text (unknown source and purpose)
- Replace permanently redirected MSDN link with its current /win32/ URL

- Whitespace changes:
    - Remove redundant end-of-line blanks (70 lines)
    - Add MarkDown indent marker compatibility spacing (14 lines)
    - Add blank line after H3 heading (1 line)

Ticket closure and reference:

Closes #4339
Closes #4470
Ref. https://github.com/MicrosoftDocs/windowsserverdocs.de-de/issues/21