ADFS/Role of Claims: Add a Claim Description URL
Description:

As reported in issue ticket #3632
(Link on "Add a Claim" referencing not-supported page.),
the link to the page points to an outdated TechNet page with permanent
redirect to the Windows Server 2012 (not even R2) page.

Thanks to Azure MVP SychevIgor for reporting & following up this issue.

Changes proposed:
- Replace outdated WS2012 page URL with slightly newer 2012R2 page URL

- Whitespace changes:
    - Remove redundant end-of-line blanks (most of the document)
    - Standardize on 1 blank line after H2 & H3 headings
    - Normalize bullet point list spacing to 1 space (2 lines)
    - Reduce spacing between metadata and page title to 1 blank line

Additional note: Remember to use Hide whitespace changes to avoid
mistaking the Diff for almost all lines being altered in content by me.

Ticket closure or reference:

Closes #3632