ADFS Overview/What's new in ADFS: PSH explained
Description:

As discussed for more than a year now in issue ticket #2514
(What is "PSH" in the context of this document?), this page needs an
explanation for the origin of the initialism "PSH", seeing that it is a
source of confusion rather than helpful reading.

Thanks to jordanmills (Jordan Mills), husamhilal, schnitz76,
jgallucci32, and ThomasKur (Thomas Kurth) for following up.

Thanks in particular to jgallucci32 for providing the correct codestyle.

Changes proposed:
- Preface the first entry of "PSH" with the full name "PowerShell"
- Place first entry of the initialism PSH in parentheses (PSH)
- Follow up last entry of PSH with the name in parentheses (PowerShell)
- Apply Camel Case: "Powershell" -> PowerShell
- Typo correction: "Brose" -> Browse
- Replace 2 old TechNet link URLs with their permanent redirect URLs

- Whitespace changes:
    - Remove redundant end-of-line blanks
    - Move HTML `</br>` codes from start of line to end of previous line
    - Bullet point list hyphen: remove single preceding blank(no indent)
    - Normalize bullet points to 1 single space between text and bullet
    - Standardize on blank line after H2/H3/H4 headings
    - Standardize on 2 blank readability lines before H2/H3/H4 headings
    - Reduce occurrences of 3 contiguous blank lines to 2
    - Add blank line between metadata and page title

Ticket closure or reference:

Closes #2514