Update What-s-New-in-DHCP.md
The DHCP Client code was updated for the recent releases; these are the most impactful changes. These are for windows client, but are values that IT admins would want to see. Sorry about the somewhat vague info ("should" be detected as metered, and "may" show up and "some Windows-based devices". Anything more crisp would be incorrect.