technical update to Feature on Demand
I have added some clarification on Feature on demand. The list "as noted" is not meant to be complete.
myself have tracked recent developements of FoD as an active user of FoD and also Windows Insider and additional to this PR, I would ask you for permission to update the table below, to make clear which FoD package added functionality.

I hope you appreciate my initiative, as now vServer Next is hitting the next LTSC insider release with b20201. 

Currently the list and other docs may appear still unappealing to IT Pros to leverage Server Core, since many restrictions existed based on 1809 do no longer exist as of today technology.