Windows Commands/Shutdown: add comment quotes
As described in the manual page document itself, "You must enclose comments in quotation marks.", those quotation marks are missing in the examples, even though the comments contain spaces between the words. This does not work when applied in a live test scenario, because the comment will not be treated as a string with spaces unless quotation marks are used.

Thanks to RodneySandwith for reporting this issue in ticket #4725 (Examples on page for Shutdown options do not show required quotes on comment parameter /c)

Also removed redundant surplus spaces in bullet point lists.
MarkDown bullet point lists only need 1 space after the bullet point list indicator, not 3 or 4 spaces.