AD FS/Deployment: URL & description corrections
From issue ticket #577
(Checklist item # 10 wrong link. Must be "Create a Claims Provider Trust"):

The links and their descriptions in the **Reference** column do not match 
the "Create a Claims Provider" step (step number 10), but instead point to 
the page about "Create a Relying Party Trust". This is clearly a mismatch.

Thanks to Jānis Veinbergs (janis-veinbergs) for reporting this issue and for
providing the correct URL/page link.

Closes #577