Remotely manage Hyper-V hosts: typos & formatting
Description:

(Ref. #712 (Managing a 2012R2 Hyper-V host from Windows 2016 Hyper-V Manager))

Although the user posting in issue ticket #712 claims that "Managing a 2012R2 
Hyper-V host from Windows 2016 Hyper-V Manager" is incorrect, I do not have 
the technical experience to refute or confirm this statement.

I can only assume that the statement "Hyper-V Manager provides the features 
available for the version of Hyper-V on the host you're managing." implies 
that newer features will not be available in the management software.

On the other hand, I have found some inconsistencies in the document, 
which I propose to rectify in this pull request.

Changes proposed:

- "Windows 2016" corrected to "Windows Server 2016" instead - 4 occurrences
- added "PowerShell" as the correct code syntax highlighting for the cmdlets
- added missing blank line after code blocks to avoid compatibility issues
- corrected the typo "lifecyle" to "lifecycle"
- added editorial blank line between the MD page index and the page title line
- added MarkDown indent marker compatibility spacing (1 space) - 5 occurrences
- added missing NewLine at end of file/document/page

Ticket closure or reference:

Ref. #712 (unless the ticket report is correct and this needs to be "Closes")