Added Chapter Performance consideration
the described experience is based on numerous customer implementations of me working as a senior IT Specialist for MS Gold Partner in relation to Windows Server, especially Windows File Server and Citrix Virtual Apps (former Xen Desktop).
Even though I am aware that extensive edits are a possible legal issue, as I am not a MS employee nor (senior) PFE, I hope you can find a balance between my PR branch and what you possibly could afford to publish.
The reason why I am requesting this - in a seperate branch - is to hopefully end the woeful discussions about ReFS is not a good filesystem on the internet community.

Imho it is and the Product Managing team should take the opportunity to include real life scenarios experience in the technical documentation, as it is vital to understand to abilities and limitations not only from a developer perspective but also from the view of implementing technology and offering the best expierence to the customers and at the end, the users.  🚀

I am willing to discuss this edit and hope it won't be declined to easily.