Windows Commands/mmc: No help at command prompt
As reported in issue ticket #3406 (Command Help /? does not function as documented), there is no available command line help for this command, at least not for any supported version of Windows or Windows Server. Maybe this command line help was available sometime 20 years ago, but it is definitely not available from Windows Server 2008 R2 onwards. It is only possible to positively affect how MMC is started with the known parameters. When using any other parameter, MMC just starts as if no parameter was given. Parameter errors are silently ignored.

Thanks to Christopher Makris (CJMakris) for reporting this documentation issue.

Proposed change:

- Remove the parameter reference /? because it is no longer used (if it ever was in use in the first place).

Closes #3406