Update add-module.md
cd \src\app would attempt to take you to <drive root>\src\app. By changing to .\src\app it goes to the src\app subdir of where you are. Assuming you just created the new tool extension, you would be in the root of the tool and .\src\app would be appropriate.