Added "rules" for 'windows-commands/dnscmd.md'
Per research results delivered by Aquent vendor Jennifer Kirsch, 2020/Dec/17. But currently the term is part of the software product that is being documented, so the documentation cannot (yet) replace the term.
This file contains the sensitive term 'slave', a term Microsoft no longer uses.

This new `"rules"` entry, for message type 'sensitive-language', prevents the OPS Build system from issuing an occurrence of the 'sensitive-language' message for this article.

Eventually, when the term is removed from the software, the term should then be removed from the dnscmd.md article file. And then the `"rules"` entry for this dnscmd.md article should also be removed, promptly.

_**CAUTION:**_ &nbsp; The `"rules"` section in docfx.json must not list any articles whose sensitive terms can all be removed now, meaning the terms can be replaced now in the documentation without misrepresenting the software product. And after an article's sensitive terms are removed, the article must be removed from the `"rules"` list promptly. Mike Jacobs and the Terminology Task Force vteam can provide further info if needed.

A loosely related sister pull request, in another repo, is:
- `https://github.com/MicrosoftDocs/sql-docs-pr/pull/18162`

Thanks.  GeneMi (= MightyPen)
. .