updating Windows Server Core limitations and FOD
this document is not intendet to be complete yet in the field often cited for the limitations of Server Core and could lead Administrators and IT decisions against the use of Server Core. 
I am adding this for the sake of completeness and later developments of Feature on Demand and hope you agree with this move.  Certainly Server 2022 is around the corner. Everywhere IT Pros, Microsoft, Vendors and Partners are vouching for the use of Server Core. 

Clear information founds better the decisions in favour for this preferred installation method of modern versions of Windows Server.

This PR is reflects the outcome of discussion with @illfated. Thanks for your advice, as always.

reference: @Karl-WE  #3016