Update robocopy.md
The /IM parameter was not documented and works differently than the /it parameter.  If you _always_ want Robocopy to overwite files in the destination you must use all three (/im /is /it) and there's forum post after forum post of people not understanding why just /is /it skipping files.  The descriptions I've added/changed here match Robocopy's own help output.