Update find.md
This is one of the command with the most ...hm.. not-totally-accurate documentation.

double quotes in the searched string are mandatory. I've added them in the syntax as they are in the help message. Also added them in the examples.

Also I don't think examples with doc files are good idea as they are not text files.

it is not true that the command does not accept wildcards in file names so I removed the part with the FOR command.

I've added the exit codes - I think it is good idea to have them for all the commands (when possible - sometimes you can't give a good description for all of them)

the example with searching all the files in a drive will be too slow and I think it's better to use a directory instead.