W32Time: Remove link to deleted KB 902229 (2003)
This PR is to resolve issue ticket #5497 (**Broken link to KB**).

The KB 902229 page is deleted, and the page can only be found in the
web archives. The article is also specific to Windows Server 2003, and
is therefore no longer relevant.

There is currently no known replacement for KB 902229, so the logical
resolution is to remove the link from both of the articles in this PR.

Proposed change:
- Remove the go.microsoft.com link redirecting to KB 902229.
- Replace the "See Also" section with the "Related topics" section
  (copied from the Windows Time Service Technical Reference page).

Additional whitespace and codestyle changes:
- Remove EOL (end-of-line) blanks (redundant whitespace).
- Standardize line spacing by adding 1 blank line after H2/H3 headings.
- Normalize bullet point lists to only one blank after bullet markers.
- Add MD indent marker spacing and **bold** to "Applies to:".
- Properly align MD tables and add centering to the values columns.

Closes #5497