Update configure-ad-fs-and-azure-mfa.md
Adding line 143 to explain that if we want to use Azure MFA as an additional option, we need to intentionally select it. The process described in the documentation will just enable it for the primary one. We got a few support tickets in Authentication pod due to the lack of this info.