Update Supported-CentOS-and-Red-Hat-Enterprise-Linux-virtual-machines-on-Hyper-V.md
Updated RHEL/CentOS 7.x grid: Row LIS Availability should indicate “built-in” in the right most column (7.0)
Updated row Windows Server 2016 Accurate Time – this feature is available in RHEL 7.4 built-in (and later) as well as with LIS 4.3. 
Updated row Live Virtual Machine Backup  that needs Note 4 in the 7.6-7.9 built-in column
Updated row Key-Value Pair needs Note 4 in the 7.6-7.9 built-in column.  
Updated row Dynamic Memory – Ballooning swhich has Note 8 in the 7.0 built-in column.  
In the Notes section, the final paragraph in Note 8 should be indented so that it is even with the text for Note 8 and Note 9.