fixing a typo
The last parameter in step 5 under " Create a VM and connect to a Virtual Network by using the Windows PowerShell Network Controller cmdlets" is wrong and is pretty hard to spot by the reader. $vmnics is used throughout that script then it changes to $vmnic in the last command. I think I already raised this issue before and I still fell for it again.