[mountvol] listing "mounts by volume" vs. "volumes by mount"
The `/l` switch lists the volumes by mount point, but there is apparently no switch to do the opposite, list mount points by volume. The information is, however, provided by typing `mountvol` without any parameters. As far as I can tell, this is the only way to get mount points by volume using `mountvol`. Unfortunately, doing it this way also shows the help info as a header, and there is no way to narrow it down to only the mount for a particular specified volume.