Update smb-known-issues.md
Title changed from "SMB known issues" to "SMB Common Issues," plus some clarifying text around the article's purpose.

The existing title makes it sound like all known SMB issues are or should be maintained here. Which is not the purpose of this article. The updates add clarity to the purpose of this article.