Add 'Additional Notes' for NTLM Block error
When the certificate replacement is executed by an account that is member of 'Protected Users', the NTLM authentication is blocked (and it's used by the replacement). This is documented in the Protected Users page, but not in the certificate replacement page. Currently, the only occurrence of the scenario is under a forum post. 
Updating here because it would have saved me 3 days of troubleshooting in a recent case, and may save others :)