Correct rndPass usage
While testing this on Windows Server 2019, I discovered that `+rndPass` shouldn't be combined with `/pass`, it is supposed to be used on it's own. Otherwise `+rndPass` becomes the actual password, which is unexpected and might not meet password complexity requirements.

Running a `ktpass /?` shows the that options `+rndPass` and `/pass` are separate from each other.