SQL Server Connectivity feature
I've been involved in migrating SUSDB from WID to a SQL Server box and we followed all steps provided, but at the end Server Manager was keeping alterting of uncompleted WSUS configuration. Though we try to run the configuration was failing due to an error resembling the folliwing entries:

2013-07-23 11:03:18  Postinstall started
2013-07-23 11:03:18  Detected role services: Api, UI, WidDatabase, Services
2013-07-23 11:03:18  Start: LoadSettingsFromXml
2013-07-23 11:03:18  Start: GetConfigValue with filename=UpdateServices-Services.xml item=ContentLocal
2013-07-23 11:03:18  Value is true
2013-07-23 11:03:18  End: GetConfigValue
2013-07-23 11:03:18  Start: GetConfigValue with filename=UpdateServices-Services.xml item=ContentDirectory
2013-07-23 11:03:18  Config file did not contain a value "ContentDirectory"
2013-07-23 11:03:18  Microsoft.UpdateServices.Administration.CommandException: A required configuration value was not found in the system. This is usually caused by installing WSUS through PowerShell and not specifying a configuration file. Review the article Managing WSUS Using PowerShell at TechNet Library (http://go.microsoft.com/fwlink/?LinkId=235499) for more information on the recommended steps to perform WSUS installation using PowerShell.
   at Microsoft.UpdateServices.Administration.PostInstall.GetConfigValue(String filename, String item)
   at Microsoft.UpdateServices.Administration.PostInstall.LoadSettingsFromXml()
   at Microsoft.UpdateServices.Administration.PostInstall.Initalize(Parameters parameters)
   at Microsoft.UpdateServices.Administration.PostInstall.Execute(String[] arguments)
Fatal Error: A required configuration value was not found in the system. This is usually caused by installing WSUS through PowerShell and not specifying a configuration file. Review the article Managing WSUS Using PowerShell at TechNet Library (http://go.microsoft.com/fwlink/?LinkId=235499) for more information on the recommended steps to perform WSUS installation using PowerShell.

We were able to resolve the issue, by removing the WID feature and installing the SQL Server Connectivity upon restart.
At that point, Server Manager was asking to complete the configuration but the this time the configuration gone smoothly.