Update remote-desktop-features.md
If mstsc is started in fullscreen mode and then switched to windowed mode during the session, the resolution is not dynamically updated and only client-side smart sizing is performed. If the client is started in windowed mode, then also only client-side smart sizing is performed ("fixed resolution" mode). Dynamic resolution change in mstsc is intended to work only when (a) client is started in fullscreen mode and (b) continues to stay in fullscreen mode