Update Install-Active-Directory-Domain-Services--Level-100-.md
Refers to ADDS installation, which is part of a series. The "Applies to" section includes 2012R2, 2016, 2019, and 2022 (2012 removed) and some paragraphs specifically mention 2012, such as "This topic explains how to install AD DS in Windows Server 2012 by using any of the following methods:". I have removed specific reference to 2012 when this applies to newer OS levels too. I have added new DomainMode and ForestMode values.