Update active-directory-functional-levels.md
I have tested this, and Windows Server 2025 DCs cannot be introduced to a domain/forest of functional level 2012 R2, an error will arise stating that they can only be introduced to forest of functional level Windows 2016 or higher. 

So current guidance is incorrect and I have modified the statement and called it out with a note.