Update Appendix-G--Securing-Administrators-Groups-in-Active-Directory.md
The document says "There should be no day-to-day user accounts in the Administrators group with the exception of the Built-in Administrator account for the domain". But it also says "Additionally, DAs and EAs inherit a number of their rights and permissions by virtue of their default membership in the Administrators group. Default group nesting for privileged groups in Active Directory should not be modified".

This is unclear and possibly contradictory. The term "day-to-day user accounts" is not defined or standard. Does it include the Administrator account and the DA and EA groups? If it does, then the part about not modifying nesting groups is contradictory. If it does not, then the Administrator account cannot be excluded, because it is not included.

It is not clear whether the intention is to remove ALL members, with the possible exception of the Administrators account; or to remove ALL members EXCEPT the default members.