Update AD-FS-Prompt-Login.md
The right command to use is Update-MgDomainFederationConfiguration not New-MgDomainFederationConfiguration. the latter creates a federation domain configuration, but in this context, this is already in existence. Therefore, an error gets returned "New-MgDomainFederationConfiguration : Domain already has Federation Configuration set."

PreferredAuthenticationProtocol - should not be changed
FederatedIdpMfaBehavior - is not always populated and shouldn't be changed in this context.