Update instructions for installing Server for NFS role
Corrected Typos of DNS when it should have been NFS.

Removed GUI instruction "1. On the **DNS Server** page, review the role description and things to note, select **Next** to continue." since it does not exist in practice. 

Corrected statement: "The installation doesn't require a reboot." In practice a reboot was indeed required/highly recommended.

~Luckystantheman